+++ src/components/DownloadAndShareComponents/downloadAndShare.types.ts
@@ -1,7 +1,5 @@
 import { z } from "zod";
 
-import { LessonItemTitle } from "../LessonItemContainer/LessonItemContainer";
-
 import {
   preselectedDownloadType,
   preselectedResourceType,
@@ -9,6 +7,7 @@ import {
   resourceFormValuesSchema,
 } from "./downloadAndShare.schema";
 
+import { LessonItemTitle } from "@/components/TeacherComponents/LessonItemContainer";
 import { ExpandingContainerTitle } from "@/components/ExpandingContainer/ExpandingContainer";
 import {
   LessonDownloadsData,
