--- src/components/ProgrammeList/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ProgrammeList";
