+++ src/components/TeacherComponents/LessonItemContainer/index.tsx
@@ -1,1 +1,2 @@
 export { LessonItemContainer } from "./LessonItemContainer";
+export type { LessonItemTitle } from "./LessonItemContainer";
