+++ src/components/TeacherComponents/SubjectProgrammeList/SubjectProgrammeList.stories.tsx
@@ -1,9 +1,10 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import AnalyticsDecorator from "../../storybook-decorators/AnalyticsDecorator";
-import { tieredProgrammeListingFixture } from "../../node-lib/curriculum-api/fixtures/tierListing.fixture";
-
 import Component from ".";
+
+import AnalyticsDecorator from "@/storybook-decorators/AnalyticsDecorator";
+import { tieredProgrammeListingFixture } from "@/node-lib/curriculum-api/fixtures/tierListing.fixture";
+
 
 export default {
   decorators: [AnalyticsDecorator],
@@ -14,8 +15,8 @@ const Template: ComponentStory<typeof Component> = (args) => (
   <Component {...args} />
 );
 
-export const ProgrammeList = Template.bind({});
+export const SubjectProgrammeList = Template.bind({});
 
-ProgrammeList.args = {
+SubjectProgrammeList.args = {
   ...tieredProgrammeListingFixture(),
 };
