+++ src/components/TeacherComponents/SubjectProgrammeListing/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./SubjectProgrammeListing";
