+++ src/pages/teachers/key-stages/[keyStageSlug]/subjects/[subjectSlug]/programmes.tsx
@@ -5,7 +5,7 @@ import curriculumApi from "@/node-lib/curriculum-api";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import AppLayout from "@/components/AppLayout/AppLayout";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
-import SubjectTierListing from "@/components/SubjectProgrammeListing/SubjectProgrammeListing";
+import SubjectTierListing from "@/components/TeacherComponents/SubjectProgrammeListing";
 import {
   getFallbackBlockingConfig,
   shouldSkipInitialBuild,
